+++ insurance-cart-microservice/src/main/java/com/example/model/AddPolicyToCartRequest.java
@@ -12,6 +12,6 @@ import lombok.ToString;
 @AllArgsConstructor
 @ToString
 public class AddPolicyToCartRequest {
-    private Integer userId;
-    private Integer policyId;
+    private Long userId;
+    private Long policyId;
 }
