+++ insurance-cart-microservice/src/main/java/com/example/model/AllItemsInCartResponse.java
@@ -10,6 +10,6 @@ import java.util.List;
 @AllArgsConstructor
 @ToString
 public class AllItemsInCartResponse {
-    private Integer userId;
-    private List<Integer> listOfPolicyIds;
+    private Long userId;
+    private List<Long> listOfPolicyIds;
 }
