+++ insurance-cart-microservice/src/main/java/com/example/model/DeletePolicyFromCartRequest.java
@@ -12,6 +12,6 @@ import lombok.ToString;
 @AllArgsConstructor
 @ToString
 public class DeletePolicyFromCartRequest {
-	private Integer userId;
-    private Integer policyId;
+    private Long userId;
+    private Long policyId;
 }
