+++ insurance-cart-microservice/src/main/java/com/example/repo/CartRepository.java
@@ -10,17 +10,17 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CartRepository extends CrudRepository<CartItems, Integer> {
+public interface CartRepository extends CrudRepository<CartItems, Long> {
     @Modifying
     @Query("DELETE FROM CartItems c WHERE c.userId = :userId AND c.policyId = :policyId")
-    public abstract void deleteByPolicyId(@Param("userId") final Integer userId, @Param("policyId") final Integer policyId);
+    public abstract void deleteByPolicyId(@Param("userId") final Long userId, @Param("policyId") final Long policyId);
 
     @Modifying
     @Query("SELECT c FROM CartItems c WHERE c.userId = :userId")
-    public abstract List<CartItems> getListOfItemsForUserId(@Param("userId") final Integer userId);
+    public abstract List<CartItems> getListOfItemsForUserId(@Param("userId") final Long userId);
 
     @Modifying
     @Query("DELETE FROM CartItems c WHERE c.userId = :userId")
-    public abstract void deleteAllByUserId(@Param("userId") final Integer userId);
+    public abstract void deleteAllByUserId(@Param("userId") final Long userId);
 
 }
